#21 Let's try provider without credentials.
